+++ samples/Catalog/configuration/Constants.ts
@@ -7,6 +7,8 @@ export const formDocumentName = 'Form_example.pdf';
 const measurementsName = 'Measurements.pdf';
 export const exampleDocumentName = 'PSPDFKit_Quickstart_Guide.pdf';
 export const examplePasswordDocumentName = 'PSPDFKit_Quickstart_Guide_Password.pdf';
+export const exampleReportName = 'JKHF-AnnualReport.pdf';
+export const exampleXFDFName = 'XFDFTest.xfdf';
 
 export const exampleImage = 'PSPDFKit_Image_Example.jpg';
 export const exampleImagePath =
@@ -30,6 +32,16 @@ export const examplePasswordDocumentPath =
       ? 'PDFs/' + examplePasswordDocumentName
       : 'file:///android_asset/' + examplePasswordDocumentName;
 
+export const exampleReportPath =
+      Platform.OS === 'ios'
+        ? 'PDFs/' + exampleReportName
+        : 'file:///android_asset/' + exampleReportName;
+
+export const exampleXFDFPath =
+        Platform.OS === 'ios'
+          ? 'PDFs/' + exampleXFDFName
+          : 'file:///android_asset/' + exampleXFDFName;
+
 export const tiffImagePath =
   Platform.OS === 'ios'
     ? 'PDFs/' + tiffImageName
@@ -51,3 +63,8 @@ export const writableFormDocumentPath =
   Platform.OS === 'ios'
     ? fileSystem.DocumentDirectoryPath + '/' + formDocumentName
     : 'file://' + fileSystem.DocumentDirectoryPath + '/' + formDocumentName;
+
+export const writableXFDFPath =
+  Platform.OS === 'ios'
+    ? fileSystem.DocumentDirectoryPath + '/' + exampleXFDFName
+    : 'file://' + fileSystem.DocumentDirectoryPath + '/' + exampleXFDFName;